requirements.txt
streamlit
python-dotenv 
pandas
selenium
beautifulsoup4
lxml 
html5lib
python-dotenv
